refactoring pw lock.
