+++ integration/src/tests/built_in_pw_lock/mod.rs
@@ -1,0 +1,2 @@
+mod build_adjust_account_transaction;
+mod build_transfer_transaction_ckb_pw;
