+++ integration/src/tests/mod.rs
@@ -3,8 +3,8 @@ pub mod build_dao_related_transaction;
 pub mod build_simple_transfer_transaction;
 pub mod build_sudt_issue_transaction;
 pub mod build_transfer_transaction_ckb;
-pub mod build_transfer_transaction_ckb_pw;
 pub mod build_transfer_transaction_udt;
+pub mod built_in_pw_lock;
 pub mod exec_instructions;
 pub mod extension_omni_lock;
 pub mod get_balance;
